Server: - Add GETKEYS: Can fetch all subkeys to a given root key in quick succession. - Fix critical bug that caused .store files to get corrupted. - Add cyon_time_us() (time in microseconds). - Change meminuse to u_int64_t.
CLI:
- Remember this cli is just a test tool and not suitable for production.
- Added test command to fill database with 1000000 test entries.

+++ src/mem.c
@@ -30,7 +30,7 @@ struct meminfo {
 	u_int16_t		magic;
 } __attribute__((__packed__));
 
-u_int32_t			meminuse;
+u_int64_t			meminuse;
 
 void
 cyon_mem_init(void)
